语义分割/语义分割/mmseg_test_rs19/mmseg_test_rs19/requirements.txt
mmcv
mmengine
mmsegmentation

